requirements.txt
spotipy>=2.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
bokeh>=3.2.0
python-dotenv>=1.0.0
scipy>=1.11.0
